Fix reset, Add auto scroll to setup.
Resets will only happen after the tab is rendered which definitely is when the user scrolls to the tab. This means that when we submit if we dont switch off the tab the data will be left there till the user tries to enter more info in which a rerender happens and everything is cleared.